+++ GameEngine/GameEngine/Transform.h
@@ -3,7 +3,7 @@
 class Transform
 {
 public:
-	float x = 0, y = 0;
+	int x = 0, y = 0;
 
 	Transform();
 	Transform(int X, int Y);
